requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Deprecated==1.2.13
Django==4.1.6
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
